Because of the way the doctrine metadata in the orm file is constructed (with the use of the entity element) it is not possible to override it easily from another project bundle.
By wrapping the information in a mapped-superclass element, it can be extended, and it tells Doctrine that the entity itself is the final entity, but a extending one. This last is the case for it extends the ProductVariant entity of Sylius self.

Doctrine documentation: https://www.doctrine-project.org/projects/doctrine-orm/en/2.6/reference/inheritance-mapping.html